+++ handlers/zenvia/zenvia.go
@@ -13,24 +13,27 @@ import (
 	"github.com/nyaruka/courier"
 	"github.com/nyaruka/courier/handlers"
 	"github.com/nyaruka/courier/utils"
+	"github.com/nyaruka/gocommon/urns"
 	"github.com/pkg/errors"
 )
 
 var (
-	maxMsgLength = 1152
-	sendURL      = "https://api-rest.zenvia.com/services/send-sms"
+	maxMsgLength    = 1152
+	whatsappSendURL = "https://api.zenvia.com/v2/channels/whatsapp/messages"
+	smsSendURL      = "https://api.zenvia.com/v2/channels/sms/messages"
 )
 
 func init() {
-	courier.RegisterHandler(newHandler())
+	courier.RegisterHandler(newHandler("ZVW", "Zenvia WhatsApp"))
+	courier.RegisterHandler(newHandler("ZVS", "Zenvia SMS"))
 }
 
 type handler struct {
 	handlers.BaseHandler
 }
 
-func newHandler() courier.ChannelHandler {
-	return &handler{handlers.NewBaseHandler(courier.ChannelType("ZV"), "Zenvia")}
+func newHandler(channelType courier.ChannelType, name string) courier.ChannelHandler {
+	return &handler{handlers.NewBaseHandler(channelType, name)}
 }
 
 // Initialize is called by the engine once everything is loaded
@@ -41,78 +44,36 @@ func (h *handler) Initialize(s courier.Server) error {
 	return nil
 }
 
-// {
-//     "callbackMoRequest": {
-// 	    	"id": "20690090",
-//         	"mobile": "555191951711",
-//         	"shortCode": "40001",
-//         	"account": "zenvia.envio",
-//         	"body": "Content of reply SMS",
-//         	"received": "2014-08-26T12:27:08.488-03:00",
-//         	"correlatedMessageSmsId": "hs765939061"
-//  	}
-// }
+type moContent struct {
+	Type         string  `json:"type"   validate:"required"`
+	Text         string  `json:"text"`
+	Payload      string  `json:"payload"`
+	FileURL      string  `json:"fileUrl"`
+	FileMimeType string  `json:"fileMimeType"`
+	FileCaption  string  `json:"fileCaption"`
+	FileName     string  `json:"fileName"`
+	Longitude    float32 `json:"longitude"`
+	Latitude     float32 `json:"latitude"`
+	Name         string  `json:"name"`
+	Address      string  `json:"address"`
+	URL          string  `json:"url"`
+}
+
 type moPayload struct {
-	CallbackMORequest struct {
-		ID         string `json:"id"                      validate:"required" `
-		From       string `json:"mobile"                  validate:"required" `
-		Text       string `json:"body"`
-		Date       string `json:"received"                validate:"required" `
-		ExternalID string `json:"correlatedMessageSmsId"`
-	} `json:"callbackMoRequest"`
-}
-
-// {
-// 		"callbackMtRequest": {
-//      	"status": "03",
-//         	"statusMessage": "Delivered",
-//         	"statusDetail": "120",
-//         	"statusDetailMessage": "Message received by mobile",
-//         	"id": "hs765939216",
-//         	"received": "2014-08-26T12:55:48.593-03:00",
-//         	"mobileOperatorName": "Claro"
-// 		}
-// }
-type statusPayload struct {
-	CallbackMTRequest struct {
-		StatusCode string `json:"status" validate:"required"`
-		ID         string `json:"id"     validate:"required" `
+	ID        string `json:"id"`
+	Timestamp string `json:"timestamp"     validate:"required"`
+	Type      string `json:"type"          validate:"required" `
+	Message   struct {
+		ID        string      `json:"id"          validate:"required"`
+		From      string      `json:"from"        validate:"required"`
+		To        string      `json:"to"          validate:"required" `
+		Direction string      `json:"direction"   validate:"required" `
+		Channel   string      `json:"channel"`
+		Contents  []moContent `json:"contents"    validate:"required" `
+	} `json:"message"`
+	Visitor struct {
+		Name string `json:"name"`
 	}
-}
-
-// {
-//     "sendSmsRequest": {
-//         "to": "555199999999",
-//         "schedule": "2014-08-22T14:55:00",
-//         "msg": "Test message.",
-//         "callbackOption": "NONE",
-//         "id": "002",
-//         "aggregateId": "1111"
-//     }
-// }
-type mtPayload struct {
-	SendSMSRequest struct {
-		To             string `json:"to"`
-		Schedule       string `json:"schedule"`
-		Msg            string `json:"msg"`
-		CallbackOption string `json:"callbackOption"`
-		ID             string `json:"id"`
-		AggregateID    string `json:"aggregateId"`
-	} `json:"sendSmsRequest"`
-}
-
-var statusMapping = map[string]courier.MsgStatusValue{
-	"00": courier.MsgSent,
-	"01": courier.MsgSent,
-	"02": courier.MsgSent,
-	"03": courier.MsgDelivered,
-	"04": courier.MsgErrored,
-	"05": courier.MsgErrored,
-	"06": courier.MsgErrored,
-	"07": courier.MsgErrored,
-	"08": courier.MsgErrored,
-	"09": courier.MsgErrored,
-	"10": courier.MsgErrored,
 }
 
 // receiveMessage is our HTTP handler function for incoming messages
@@ -124,23 +85,75 @@ func (h *handler) receiveMessage(ctx context.Context, channel courier.Channel, w
 		return nil, handlers.WriteAndLogRequestError(ctx, h, channel, w, r, err)
 	}
 
+	if strings.ToUpper(payload.Type) != "MESSAGE" {
+		return nil, handlers.WriteAndLogRequestError(ctx, h, channel, w, r, fmt.Errorf("unsupported event type: %s", payload.Type))
+	}
+
 	// create our date from the timestamp
-	// 2017-05-03T06:04:45.345-03:00
-	date, err := time.Parse("2006-01-02T15:04:05.000-07:00", payload.CallbackMORequest.Date)
+	// 2017-05-03T06:04:45Z
+	date, err := time.Parse("2006-01-02T15:04:05Z", payload.Timestamp)
 	if err != nil {
-		return nil, handlers.WriteAndLogRequestError(ctx, h, channel, w, r, fmt.Errorf("invalid date format: %s", payload.CallbackMORequest.Date))
+		return nil, handlers.WriteAndLogRequestError(ctx, h, channel, w, r, fmt.Errorf("invalid date format: %s", payload.Timestamp))
+	}
+
+	if strings.ToUpper(payload.Message.Direction) != "IN" {
+		return nil, handlers.WriteAndLogRequestIgnored(ctx, h, channel, w, r, "ignoring request, not incoming messages")
 	}
 
 	// create our URN
-	urn, err := handlers.StrictTelForCountry(payload.CallbackMORequest.From, channel.Country())
+	urn, err := urns.NewWhatsAppURN(payload.Message.From)
 	if err != nil {
 		return nil, handlers.WriteAndLogRequestError(ctx, h, channel, w, r, err)
 	}
 
-	// build our msg
-	msg := h.Backend().NewIncomingMsg(channel, urn, payload.CallbackMORequest.Text).WithExternalID(payload.CallbackMORequest.ID).WithReceivedOn(date.UTC())
-	// and finally write our message
-	return handlers.WriteMsgsAndResponse(ctx, h, []courier.Msg{msg}, w, r)
+	contactName := payload.Visitor.Name
+
+	msgs := []courier.Msg{}
+
+	for _, content := range payload.Message.Contents {
+
+		text := ""
+		mediaURL := ""
+
+		if content.Type == "text" {
+			text = content.Text
+		} else if content.Type == "location" {
+			mediaURL = fmt.Sprintf("geo:%f,%f", content.Latitude, content.Longitude)
+		} else if content.Type == "file" {
+			mediaURL = content.FileURL
+		} else {
+			// we received a message type we do not support.
+			courier.LogRequestError(r, channel, fmt.Errorf("unsupported message type %s", content.Type))
+		}
+
+		// build our msg
+		msg := h.Backend().NewIncomingMsg(channel, urn, text).WithExternalID(payload.Message.ID).WithReceivedOn(date.UTC()).WithContactName(contactName)
+		if mediaURL != "" {
+			msg.WithAttachment(mediaURL)
+		}
+		msgs = append(msgs, msg)
+	}
+
+	// and finally write our messages
+	return handlers.WriteMsgsAndResponse(ctx, h, msgs, w, r)
+}
+
+var statusMapping = map[string]courier.MsgStatusValue{
+	"REJECTED":      courier.MsgFailed,
+	"NOT_DELIVERED": courier.MsgFailed,
+	"SENT":          courier.MsgSent,
+	"DELIVERED":     courier.MsgDelivered,
+	"READ":          courier.MsgDelivered,
+}
+
+type statusPayload struct {
+	ID            string `json:"id"`
+	Type          string `json:"type"       validate:"required" `
+	MessageID     string `json:"messageId"`
+	MessageStatus struct {
+		Timestamp string `json:"timestamp"`
+		Code      string `json:"code"`
+	} `json:"messageStatus"`
 }
 
 // receiveStatus is our HTTP handler function for status updates
@@ -152,69 +165,118 @@ func (h *handler) receiveStatus(ctx context.Context, channel courier.Channel, w 
 		return nil, handlers.WriteAndLogRequestError(ctx, h, channel, w, r, err)
 	}
 
-	msgStatus, found := statusMapping[payload.CallbackMTRequest.StatusCode]
+	if strings.ToUpper(payload.Type) != "MESSAGE_STATUS" {
+		return nil, handlers.WriteAndLogRequestError(ctx, h, channel, w, r, fmt.Errorf("unsupported event type: %s", payload.Type))
+	}
+
+	msgStatus, found := statusMapping[strings.ToUpper(payload.MessageStatus.Code)]
 	if !found {
 		msgStatus = courier.MsgErrored
 	}
 
 	// write our status
-	status := h.Backend().NewMsgStatusForExternalID(channel, payload.CallbackMTRequest.ID, msgStatus)
+	status := h.Backend().NewMsgStatusForExternalID(channel, payload.MessageID, msgStatus)
 	return handlers.WriteMsgStatusAndResponse(ctx, h, channel, status, w, r)
 
 }
 
+//
+type mtContent struct {
+	Type         string `json:"type"`
+	Text         string `json:"text,omitempty"`
+	FileURL      string `json:"fileUrl,omitempty"`
+	FileMimeType string `json:"fileMimeType,omitempty"`
+	FileCaption  string `json:"fileCaption,omitempty"`
+	FileName     string `json:"fileName,omitempty"`
+}
+
+type mtPayload struct {
+	From     string      `json:"from"`
+	To       string      `json:"to"`
+	Contents []mtContent `json:"contents"`
+}
+
 // SendMsg sends the passed in message, returning any error
 func (h *handler) SendMsg(ctx context.Context, msg courier.Msg) (courier.MsgStatus, error) {
-	username := msg.Channel().StringConfigForKey(courier.ConfigUsername, "")
-	if username == "" {
-		return nil, fmt.Errorf("no username set for ZV channel")
+	channel := msg.Channel()
+
+	token := channel.StringConfigForKey(courier.ConfigAPIKey, "")
+	if token == "" {
+		return nil, fmt.Errorf("no token set for ZVW channel")
 	}
 
-	password := msg.Channel().StringConfigForKey(courier.ConfigPassword, "")
-	if password == "" {
-		return nil, fmt.Errorf("no password set for ZV channel")
+	payload := mtPayload{
+		From: strings.TrimLeft(channel.Address(), "+"),
+		To:   strings.TrimLeft(msg.URN().Path(), "+"),
 	}
 
-	status := h.Backend().NewMsgStatusForID(msg.Channel(), msg.ID(), courier.MsgErrored)
-	parts := handlers.SplitMsgByChannel(msg.Channel(), handlers.GetTextAndAttachments(msg), maxMsgLength)
-	for _, part := range parts {
-		zvMsg := mtPayload{}
-		zvMsg.SendSMSRequest.To = strings.TrimLeft(msg.URN().Path(), "+")
-		zvMsg.SendSMSRequest.Msg = part
-		zvMsg.SendSMSRequest.ID = msg.ID().String()
-		zvMsg.SendSMSRequest.CallbackOption = "FINAL"
+	status := h.Backend().NewMsgStatusForID(channel, msg.ID(), courier.MsgErrored)
 
-		requestBody := new(bytes.Buffer)
-		json.NewEncoder(requestBody).Encode(zvMsg)
+	text := ""
+	if channel.ChannelType() == "ZVW" {
+		for _, attachment := range msg.Attachments() {
+			attType, attURL := handlers.SplitAttachment(attachment)
+			payload.Contents = append(payload.Contents, mtContent{
+				Type:         "file",
+				FileURL:      attURL,
+				FileMimeType: attType,
+			})
 
-		// build our request
-		req, err := http.NewRequest(http.MethodPost, sendURL, requestBody)
-		if err != nil {
-			return nil, err
 		}
-		req.Header.Set("Content-Type", "application/json")
-		req.Header.Set("Accept", "application/json")
-		req.SetBasicAuth(username, password)
+		text = msg.Text()
 
-		rr, err := utils.MakeHTTPRequest(req)
-
-		// record our status and log
-		log := courier.NewChannelLogFromRR("Message Sent", msg.Channel(), msg.ID(), rr).WithError("Message Send Error", err)
-		status.AddLog(log)
-		if err != nil {
-			return status, nil
-		}
-
-		// was this request successful?
-		responseMsgStatus, _ := jsonparser.GetString(rr.Body, "sendSmsResponse", "statusCode")
-		msgStatus, found := statusMapping[responseMsgStatus]
-		if msgStatus == courier.MsgErrored || !found {
-			log.WithError("Message Send Error", errors.Errorf("received non-success response from Zenvia '%s'", responseMsgStatus))
-			return status, nil
-		}
-
-		status.SetStatus(courier.MsgWired)
+	} else if channel.ChannelType() == "ZVS" {
+		text = handlers.GetTextAndAttachments(msg)
 	}
+
+	msgParts := make([]string, 0)
+	if text != "" {
+		msgParts = handlers.SplitMsgByChannel(channel, text, maxMsgLength)
+	}
+
+	for _, msgPart := range msgParts {
+		payload.Contents = append(payload.Contents, mtContent{
+			Type: "text",
+			Text: msgPart,
+		})
+	}
+
+	jsonBody, err := json.Marshal(payload)
+	if err != nil {
+		return status, err
+	}
+
+	sendURL := whatsappSendURL
+	if channel.ChannelType() == "ZVS" {
+		sendURL = smsSendURL
+	}
+
+	req, err := http.NewRequest(http.MethodPost, sendURL, bytes.NewReader(jsonBody))
+
+	if err != nil {
+		return nil, err
+	}
+	req.Header.Set("Content-Type", "application/json")
+	req.Header.Set("Accept", "application/json")
+	req.Header.Set("X-API-TOKEN", token)
+
+	rr, err := utils.MakeHTTPRequest(req)
+
+	// record our status and log
+	log := courier.NewChannelLogFromRR("Message Sent", msg.Channel(), msg.ID(), rr).WithError("Message Send Error", err)
+	status.AddLog(log)
+	if err != nil {
+		return status, nil
+	}
+
+	externalID, err := jsonparser.GetString(rr.Body, "id")
+	if err != nil {
+		log.WithError("Message Send Error", errors.Errorf("unable to get id from body"))
+		return status, nil
+	}
+
+	status.SetExternalID(externalID)
+	// this was wired successfully
+	status.SetStatus(courier.MsgWired)
 	return status, nil
-
 }
